Avances del modulo de planes

+++ app/Http/Requests/StorePlan.php
@@ -36,7 +36,7 @@ class StorePlan extends FormRequest
     {
         return [
             'name' => ['required', 'string'],
-            'price' => ['required', 'double']
+            'price' => ['required', 'numeric']
         ];
     }
 }
